Move MediaPlayObserver and TTSObserver to main package
