+++ app/src/main/java/com/grammatek/simaromur/AppRepository.java
@@ -13,9 +13,7 @@ import com.grammatek.simaromur.db.AppDataDao;
 import com.grammatek.simaromur.db.ApplicationDb;
 import com.grammatek.simaromur.db.Voice;
 import com.grammatek.simaromur.db.VoiceDao;
-import com.grammatek.simaromur.network.tiro.MediaPlayObserver;
 import com.grammatek.simaromur.network.tiro.SpeakController;
-import com.grammatek.simaromur.network.tiro.TTSObserver;
 import com.grammatek.simaromur.network.tiro.VoiceController;
 import com.grammatek.simaromur.network.tiro.pojo.SpeakRequest;
 import com.grammatek.simaromur.network.tiro.pojo.VoiceResponse;
